surface residue integration with main
